Feat/automatic validation

+++ src/main/java/com/safesteps/backend/domain/users/dto/UserResponseDTO.java
@@ -17,7 +17,7 @@ public class UserResponseDTO {
     private Integer points;
     private Long level;
     private Boolean isAnonymous;
-    private Integer reputacio;
+    private double reputacio;
     private OffsetDateTime createdAt;
 
     public UserResponseDTO(User user) {
